Cart footer - code refactor
