+++ src/pages/fileCentricCart/cartView.js
@@ -1,13 +1,12 @@
 import React from 'react';
 import {
   Grid, withStyles,
-  IconButton, Link,
+  IconButton,
 } from '@material-ui/core';
-import TextField from '@material-ui/core/TextField';
 import { DeleteOutline as DeleteOutlineIcon, ArrowDropDown as ArrowDropDownIcon } from '@material-ui/icons';
-import { ToolTip as Tooltip } from 'bento-components';
 import CartBody from './components/body/cartBody';
 import CartHeader from './components/header/cartHeader';
+import CartFooter from './components/footer/cartFooter';
 import DialogBox from './components/dialogBox/dialogBox';
 import Styles from './cartView.style';
 import client from '../../utils/graphqlClient';
@@ -133,18 +132,6 @@ const cartView = ({
 
   const numberOfFilesBeDeleted = myFilesPageData.popUpWindow.showNumberOfFileBeRemoved ? fileIDs.length : '';
 
-  const toolTipIcon = ({ title, placement }) => (
-    <Tooltip arrow title={title} placement={placement}>
-      <IconButton className={classes.helpBtn} aria-label="help">
-        <img
-          src={myFilesPageData.tooltipIcon}
-          alt={myFilesPageData.tooltipAlt}
-          className={classes.helpIcon}
-        />
-      </IconButton>
-    </Tooltip>
-  );
-
   return (
     <Grid className={classes.marginTopNegative20}>
       {/* Section: DialogBox */}
@@ -181,56 +168,13 @@ const cartView = ({
               tableDownloadCSV={tableDownloadCSV}
             />
 
-            {/* Section: Bottom controls */}
-            <div className={classes.paddingLeftRight}>
-              <div className={classes.message}>
-                <span>
-                  To access and analyze files: select and remove unwanted files,
-                  click the “Download Manifest” button, and upload the resulting
-                  Manifest file to your
-                  {' '}
-                  <Link target="_blank" className={classes.link} href="http://www.cancergenomicscloud.org/">
-                    Seven Bridges Genomics
-                  </Link>
-                  <img
-                    src={externalLinkIcon.src}
-                    alt={externalLinkIcon.alt}
-                    className={classes.linkIcon}
-                  />
-                  {' '}
-                  account.
-                </span>
-              </div>
-              {/* Section: User Comments */}
-              <div className={classes.manifestTextarea}>
-                <TextField
-                  id="multiline-user-coments"
-                  label={myFilesPageData.textareaPlaceholder}
-                  multiline
-                  rows={6}
-                  style={{ minWidth: '550px' }}
-                  className={classes.textField}
-                  margin="dense"
-                  variant="filled"
-                  onChange={(e) => setUserComments(e.target.value)}
-                />
-                {toolTipIcon({ title: myFilesPageData.userCommentsTooltipMessage, placement: 'right' })}
-              </div>
-
-              {/* Section: Button Group */}
-              <div className={classes.buttonGroup}>
-                <button
-                  type="button"
-                  className={classes.downloadButton}
-                  onClick={() => prepareDownload()}
-                >
-                  {myFilesPageData.downButtonText}
-                  {' '}
-                </button>
-                {toolTipIcon({ title: myFilesPageData.downloadBtnTooltipMessage, placement: 'right' })}
-              </div>
-            </div>
-
+            {/* Section: Footer */}
+            <CartFooter
+              myFilesPageData={myFilesPageData}
+              externalLinkIcon={externalLinkIcon}
+              setUserComment={(e) => setUserComments(e.target.value)}
+              preparedownload={() => prepareDownload()}
+            />
           </div>
         </div>
       </Grid>
