+++ src/pages/fileCentricCart/cartView.style.js
@@ -32,17 +32,6 @@ export default (theme) => ({
       color: '#dc762f',
     },
   },
-  linkIcon: {
-    color: '#dc762f',
-    width: '20px',
-    verticalAlign: 'sub',
-    margin: '0px 0px 0px 2px',
-  },
-  helpIcon: {
-    verticalAlign: 'top',
-    width: '17px',
-    zIndex: '600',
-  },
   topButtonGroup: {
     textAlign: 'right',
     padding: '10px 43px 15px 0px',
@@ -63,31 +52,6 @@ export default (theme) => ({
   manifestButtonGroup: {
     marginTop: '10px',
     float: 'right',
-  },
-  manifestTextarea: {
-    position: 'relative',
-    marginTop: '20px',
-  },
-  downloadButton: {
-    height: '36px',
-    minWidth: '191px',
-    color: '#fff',
-    boxShadow: 'none',
-    backgroundColor: '#3890c5',
-    padding: '6px 16px',
-    fontSize: '0.875rem',
-    boxSizing: 'border-box',
-    transition: 'background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
-    lineHeight: '1.75',
-    fontWeight: '500',
-    fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
-    borderRadius: '4px',
-    textTransform: 'uppercase',
-    border: 'none',
-    verticalAlign: 'top',
-    marginTop: '6px',
-    marginBottom: '19px',
-    cursor: 'pointer',
   },
   popUpWindowText: {
     fontFamily: 'Lato',
