+++ src/pages/fileCentricCart/components/footer/cartFooter.style.js
@@ -1,3 +1,61 @@
 export default () => ({
-
+  paddingLeftRight: {
+    paddingLeft: '30px',
+    paddingRight: '30px',
+  },
+  message: {
+    color: '#000000',
+    fontSize: '15px',
+    fontFamily: '"Open Sans", sans-serif',
+    lineHeight: '22px',
+    width: '35%',
+    marginBottom: '5px',
+    paddingLeft: '7px',
+  },
+  helpBtn: {
+    padding: '0px',
+  },
+  link: {
+    color: '#dc762f',
+    fontWeight: 'bolder',
+    '&:hover': {
+      color: '#dc762f',
+    },
+  },
+  linkIcon: {
+    color: '#dc762f',
+    width: '20px',
+    verticalAlign: 'sub',
+    margin: '0px 0px 0px 2px',
+  },
+  manifestTextarea: {
+    position: 'relative',
+    marginTop: '20px',
+  },
+  downloadButton: {
+    height: '36px',
+    minWidth: '191px',
+    color: '#fff',
+    boxShadow: 'none',
+    backgroundColor: '#3890c5',
+    padding: '6px 16px',
+    fontSize: '0.875rem',
+    boxSizing: 'border-box',
+    transition: 'background-color 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,box-shadow 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms,border 250ms cubic-bezier(0.4, 0, 0.2, 1) 0ms',
+    lineHeight: '1.75',
+    fontWeight: '500',
+    fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
+    borderRadius: '4px',
+    textTransform: 'uppercase',
+    border: 'none',
+    verticalAlign: 'top',
+    marginTop: '6px',
+    marginBottom: '19px',
+    cursor: 'pointer',
+  },
+  helpIcon: {
+    verticalAlign: 'top',
+    width: '17px',
+    zIndex: '600',
+  },
 });
